refactor: project 컴포넌트를 올바른 위치(shared/)로 복구 및 queryFn 추가
변경 사항:
- features/project-detail/ → shared/components/project/로 이동
  - ProjectClient: Shell 컴포넌트 (모든 뷰의 공통 래퍼)
  - ProjectHeader: 공통 헤더
  - ViewNavigator: 뷰 전환 탭
  - TaskDetailPanel: 작업 상세 패널 (WBS에서 사용)
- ProjectClient에 queryFn 추가 (재검증 지원)
- task_id 속성 제거 (StoredProject 인터페이스에 없음)
- 사용되지 않는 파일 삭제: ProjectView.tsx, ViewSelector.tsx
- 모든 import 경로 업데이트

아키텍처 복구 근거:
- project 컴포넌트들은 Shell(공통 래퍼) 역할
- 모든 뷰(wbs/gantt/kanban)가 공유하는 레이아웃
- features/는 독립적인 View 컴포넌트만 포함

queryFn 필수 이유:
- HydrationBoundary는 초기 데이터만 제공
- 재검증 시(window focus, staleTime 만료) queryFn 필요
- useProjectQuery 훅 사용으로 일관성 확보
