+++ shared/components/project/ViewNavigator.tsx
