+++ shared/lib/queries/projectService.ts
@@ -24,7 +24,6 @@ export const getProjectById = cache(async (projectId: string) => {
       description: storedProject.description,
       teamSize: storedProject.teamSize,
       duration: storedProject.duration,
-      task_id: storedProject.task_id,
     };
   }
 
@@ -37,7 +36,6 @@ export const getProjectById = cache(async (projectId: string) => {
       description: mockProject.description,
       teamSize: mockProject.teamMembers.length,
       duration: 90,
-      task_id: projectId,
     };
   }
